+++ src/language/parse/index.ts
@@ -1,0 +1,1 @@
+export { parse } from "./parse";
